move association into patient/profile
