+++ apps/modernization-api/src/main/java/gov/cdc/nbs/patient/profile/association/AssociatedWith.java
@@ -1,4 +1,4 @@
-package gov.cdc.nbs.event.search.investigation.association;
+package gov.cdc.nbs.patient.profile.association;
 
 public record AssociatedWith(
         long id,
