+++ apps/modernization-api/src/main/java/gov/cdc/nbs/patient/profile/association/AssociatedWithTupleMapper.java
@@ -1,4 +1,4 @@
-package gov.cdc.nbs.event.search.investigation.association;
+package gov.cdc.nbs.patient.profile.association;
 
 import com.querydsl.core.Tuple;
 import gov.cdc.nbs.entity.odse.QPublicHealthCase;
