+++ apps/modernization-api/src/main/java/gov/cdc/nbs/patient/profile/vaccination/PatientVaccinationTupleMapper.java
@@ -6,8 +6,8 @@ import gov.cdc.nbs.entity.odse.QPersonName;
 import gov.cdc.nbs.entity.odse.QPublicHealthCase;
 import gov.cdc.nbs.entity.srte.QCodeValueGeneral;
 import gov.cdc.nbs.entity.srte.QConditionCode;
-import gov.cdc.nbs.event.search.investigation.association.AssociatedWith;
-import gov.cdc.nbs.event.search.investigation.association.AssociatedWithTupleMapper;
+import gov.cdc.nbs.patient.profile.association.AssociatedWith;
+import gov.cdc.nbs.patient.profile.association.AssociatedWithTupleMapper;
 import gov.cdc.nbs.provider.ProviderNameTupleMapper;
 
 import java.time.Instant;
