+++ apps/modernization-api/src/test/java/gov/cdc/nbs/patient/contact/PatientNamedByContactTupleMapperTest.java
@@ -1,7 +1,7 @@
 package gov.cdc.nbs.patient.contact;
 
 import com.querydsl.core.Tuple;
-import gov.cdc.nbs.event.search.investigation.association.AssociatedWith;
+import gov.cdc.nbs.patient.profile.association.AssociatedWith;
 import org.junit.jupiter.api.Test;
 
 import java.time.Instant;
